refactor(CommendRenderer): domain과 view 분리
