+++ src/main/java/chess/controller/commend/CommendType.java
@@ -1,0 +1,7 @@
+package chess.controller.commend;
+
+public enum CommendType {
+    START,
+    END,
+    MOVE;
+}
