+++ src/main/java/chess/controller/commend/StartCommend.java
@@ -1,7 +1,8 @@
-package chess.view.commend;
+package chess.controller.commend;
 
 import chess.controller.ChessController;
 import chess.domain.ChessGame;
+import chess.renderer.CommendRenderer;
 import chess.view.validator.ValidateType;
 
 import java.util.List;
@@ -13,7 +14,7 @@ public class StartCommend extends Commend {
 
     @Override
     public boolean operate(ChessGame chessGame) {
-        if (CommendType.match(inputView.requestCommend(List.of(ValidateType.START)).get(0)).equals(CommendType.END)) {
+        if (CommendRenderer.render(inputView.requestCommend(List.of(ValidateType.START)).get(0)).equals(CommendType.END)) {
             this.chessController.setCommend(new EndCommend(chessController));
             return true;
         }
