+++ src/main/java/chess/view/validator/NotInRangeValidator.java
@@ -3,8 +3,8 @@ package chess.view.validator;
 import java.util.regex.Pattern;
 
 public class NotInRangeValidator implements InputValidator {
-    private static final Pattern FILE_RANGE = Pattern.compile("^[a-f]");
-    private static final Pattern RANK_RANGE = Pattern.compile("^[1-8]");
+    private static final Pattern FILE_RANGE = Pattern.compile("[^a-f]");
+    private static final Pattern RANK_RANGE = Pattern.compile("[^1-8]");
     private static final String OUT_OF_RANGE_MASSAGE = "범위에서 벗어났습니다.";
     private InputValidator next;
